Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 int main ()
 {
-    float a , b ,c ;
+    int a , b ,c ;
     float average;
     float sum ;
     sum = a + b + c ;
     average = sum / 3 ;
 
-    scanf("%f %f %f", &a ,&b ,&c);
-    printf("Average:%d",average);
+    scanf("%d %d %d", &a ,&b ,&c);
+    printf("Average:%.2f",average);
     return 0;
 
 }
